pyproject.toml
[tool.poetry]
name = "slide-viewer"
version = "0.1.0"
description = "View for slide scans."
authors = ["Luke Shiner <luke@lukeshiner.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
gunicorn = "^21.2.0"
django-admin-sortable2 = "^2.1.10"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"
black = "^24.3.0"
isort = "^5.13.2"
djlint = "^1.34.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
